Adjust trailer handling in grpc-web call
Trailers are now polled as a frame in the body - so we have to handle them at the same time that poll data.

If trailers are encountered during client decode, they are stored and returned once poll_frame stops returning data frames.

Since GRPC web frames are not aligned with underlying HTTP frames, we can’t simply return trailers once we find them - we have to use the presence of trailers to indicate the end of message frames.