Work around Cargo Issue 10788
See https://github.com/rust-lang/cargo/issues/10788 for the full issue.

When enabling a feature of a dependency, the syntax `tokio/rt` implicitly required a feature `tokio` to exist in 1.70 of cargo. By 1.77 this has been fixed. To work around this, we can use "weak feature dependencies" like `tokio?/rt` which enables the `rt` feature of tokio only if tokio is itself an enabled dependency. Also adding `dep:tokio` to the list of dependencies ensures that this is always true.

When the MSRV is upgraded again, we can probably remove the weak `?` and the `dep:tokio` in these cases.